Trim skill registry module size
Move the staged commit error type into the staged-install\nmodule so src/skills/registry.rs stays under the project\nfile-size limit while preserving the existing public API via\nre-export.\n\nAlso replace the remaining unwrap in the empty-directory\nregistry test with a descriptive expect message as requested\nby review.
