+++ src/skills/registry.rs
@@ -20,7 +20,7 @@ use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 
 pub use loading::{check_gating, compute_hash};
-pub use staged_install::{PreparedSkillInstall, SkillInstallPayload};
+pub use staged_install::{CommitPreparedInstallError, PreparedSkillInstall, SkillInstallPayload};
 
 use crate::skills::bundle::SkillBundleError;
 use crate::skills::{LoadedSkill, SkillSource, SkillTrust};
@@ -77,49 +77,6 @@ pub enum SkillRegistryError {
 
     #[error("Invalid skill content: {reason}")]
     InvalidContent { reason: String },
-}
-
-/// A failed staged install commit that preserves the prepared filesystem state.
-///
-/// [`SkillRegistry::commit_install`] returns this when the final duplicate
-/// check or rename step fails after preparation has already produced a
-/// [`PreparedSkillInstall`]. The original [`SkillRegistryError`] is preserved
-/// in `error`, and `prepared` is returned so callers can inspect or roll back
-/// the staged directory with [`SkillRegistry::cleanup_prepared_install`]
-/// without reconstructing install state.
-#[derive(Debug)]
-pub struct CommitPreparedInstallError {
-    error: SkillRegistryError,
-    prepared: Box<PreparedSkillInstall>,
-}
-
-impl CommitPreparedInstallError {
-    /// Borrow the original registry error that caused the commit to fail.
-    pub fn error(&self) -> &SkillRegistryError {
-        &self.error
-    }
-
-    /// Borrow the prepared install state that remained staged after failure.
-    pub fn prepared(&self) -> &PreparedSkillInstall {
-        &self.prepared
-    }
-
-    /// Split the failure into its original error and prepared install state.
-    pub fn into_parts(self) -> (SkillRegistryError, PreparedSkillInstall) {
-        (self.error, *self.prepared)
-    }
-}
-
-impl std::fmt::Display for CommitPreparedInstallError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        self.error.fmt(f)
-    }
-}
-
-impl std::error::Error for CommitPreparedInstallError {
-    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
-        Some(&self.error)
-    }
 }
 
 /// Registry of available skills.
